Fix drive_diff

+++ robot2/mechanics.h
@@ -8,6 +8,6 @@
 // left_delay: The microsecond delay for the left steppers.
 // right_delay: The microsecnd delay for the right steppers.
 // nsteps: The maximum number of steps a motor will take.
-void drive_diff(int left_delay, int right_delay, int nsteps);
+void drive_diff(unsigned long left_delay, unsigned long right_delay, int nsteps);
 
 #endif
